fix(neoforge): re-derive mixin companions for the 26.2 GUI and entity changes
Minecraft 26.2 reshaped several vanilla targets, and NeoForge patches some of
them into different shapes than Fabric, so the loader-specific companions are
re-derived against the 26.2 NeoForge-patched jar:

- Gui -> Hud split: the HUD hooks (overlay start, exp-level tweak) moved to the
  new Hud class. The shared mixin is renamed accordingly and the companions
  (MixinGui -> MixinHud) now target Hud; NeoForge still splits the hotbar and
  experience extraction, so it keeps its own anchors.
- Screen render: upstream now fires ScreenRenderEvent from Gui#extractRenderState.
  NeoForge routes that screen extraction through ClientHooks#extractScreen, so
  the hook is split out of the shared client MixinGui into per-loader companions
  (Fabric anchors the vanilla Screen#extractRenderStateWithTooltipAndSubtitles
  call, NeoForge the ClientHooks#extractScreen call). This replaces the old
  GameRenderer#extractGui companions, which upstream removed.
- Slime bounce: NeoForge moves Entity#restituteMovementAfterCollisions into an
  added BlockPos overload, so the Math#max hook is split out of the shared
  MixinEntity into per-loader companions.

Both loaders boot to the LiquidBounce UI on Minecraft 26.2 with no injection
errors; an offline descriptor-exact scan of the shared mixins against the 26.2
NeoForge-patched jar reports only the known lambda false positive.
