+++ neoforge/src/main/java/net/ccbluex/liquidbounce/injection/mixins/neoforge/MixinGui.java
@@ -18,50 +18,40 @@
  */
 package net.ccbluex.liquidbounce.injection.mixins.neoforge;
 
-import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
-import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
-import net.ccbluex.liquidbounce.integration.theme.component.HudComponentManager;
-import net.ccbluex.liquidbounce.integration.theme.component.HudComponentTweak;
-import net.ccbluex.liquidbounce.interfaces.GuiAddition;
+import com.llamalad7.mixinextras.sugar.Local;
+import net.ccbluex.liquidbounce.event.EventManager;
+import net.ccbluex.liquidbounce.event.events.ScreenRenderEvent;
 import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.gui.Gui;
 import net.minecraft.client.gui.GuiGraphicsExtractor;
-import net.minecraft.client.multiplayer.MultiPlayerGameMode;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 /**
- * Loader-specific companion of {@code minecraft.gui.MixinGui}.
+ * Loader-specific companion of {@code minecraft.client.MixinGui}.
  * <p>
- * NeoForge splits {@code extractHotbarAndDecorations} into individual GUI
- * layer methods, so each loader anchors these hooks in its own shape of the
- * hotbar and experience extraction.
+ * NeoForge patches the screen extraction in {@code extractRenderState} through
+ * {@code ClientHooks#extractScreen}, so the screen render event is anchored
+ * after that call. The {@code GuiGraphicsExtractor} is captured by type because
+ * the recompiled NeoForge jar does not preserve local variable names.
  */
 @Mixin(Gui.class)
 public abstract class MixinGui {
 
     /**
-     * Hook render hud event at the top layer.
-     * <p>
-     * {@code extractHotbar} handles the spectator branch internally, making its
-     * head equivalent to the head of the vanilla {@code extractHotbarAndDecorations}.
+     * Hook screen render event
      */
-    @Inject(method = "extractHotbar", at = @At("HEAD"))
-    private void hookRenderEventStart(GuiGraphicsExtractor context, DeltaTracker tickCounter, CallbackInfo ci) {
-        ((GuiAddition) this).liquid_bounce$extractOverlay(context, tickCounter);
-    }
-
-    @WrapOperation(
-        method = "extractExperienceLevel",
-        at = @At(value = "INVOKE", target = "Lnet/minecraft/client/multiplayer/MultiPlayerGameMode;hasExperience()Z")
-    )
-    private boolean tweakExpLevelText(MultiPlayerGameMode instance, Operation<Boolean> original) {
-        if (HudComponentManager.isTweakEnabled(HudComponentTweak.DISABLE_EXP_BAR)) {
-            return false;
-        }
-        return original.call(instance);
+    @Inject(method = "extractRenderState", at = @At(value = "INVOKE",
+        target = "Lnet/neoforged/neoforge/client/ClientHooks;extractScreen("
+            + "Lnet/minecraft/client/gui/screens/Screen;Ljava/util/Stack;"
+            + "Lnet/minecraft/client/gui/GuiGraphicsExtractor;IIF)V",
+        remap = false,
+        shift = At.Shift.AFTER))
+    public void hookScreenRender(DeltaTracker deltaTracker, boolean shouldRenderLevel, boolean resourcesLoaded,
+        CallbackInfo ci, @Local GuiGraphicsExtractor graphics) {
+        EventManager.INSTANCE.callEvent(new ScreenRenderEvent(graphics, deltaTracker.getGameTimeDeltaPartialTick(false)));
     }
 
 }
